backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
boto3
python-dotenv
pdfplumber
chromadb
sentence-transformers
google-generativeai
requests
pydantic
python-multipart
